+++ lib/activejob/temporal/workflows/aj_workflow.rb
@@ -8,6 +8,7 @@ require "temporalio/workflow"
 
 require_relative "../activities/rate_limit_activity"
 require_relative "dead_letter_support"
+require_relative "workflow_child_workflows"
 require_relative "workflow_chaining"
 require_relative "workflow_dependencies"
 require_relative "workflow_interactions"
@@ -49,6 +50,7 @@ module ActiveJob
       # @see https://docs.temporal.io/workflows#timers Temporal Durable Timers
       class AjWorkflow < Temporalio::Workflow::Definition
         include DeadLetterSupport
+        include WorkflowChildWorkflows
         include WorkflowChaining
         include WorkflowDependencies
         include WorkflowInteractions
@@ -143,7 +145,9 @@ module ActiveJob
 
           wait_until_scheduled(payload)
           wait_for_dependencies(payload)
-          result = execute_activity_sequence(payload) do |chain_payload|
+          result = execute_activity_payload(payload)
+          result = execute_child_workflow_sequence(payload, result)
+          result = execute_chain_sequence(payload, result) do |chain_payload|
             current_activity_payload = chain_payload
           end